ActiveRecord no longer has a "scoped" method
Using the newer "none" method is a great way to do what the comment said was being done, return an AR Relation that's a null set.

I'm not totally sure why this code is doing that or what's going on in the ActiveRecord::Wrapper class around 'sets' in general.

I was unable to figure out how to make a test case that exersizes this code.

Closes #49