oai-pmh identifiers use colon to separate namespace and id, not slash
This is in conformance with oai-pmh non-mandatory guidelines: http://www.openarchives.org/OAI/2.0/guidelines-oai-identifier.htm

It also matches how sample identifier in 'Identify' was being generated before.
https://github.com/code4lib/ruby-oai/blob/23e3ac83c43231fbf55f8dfb70175dbe0975b914/lib/oai/provider/response/identify.rb#L24

It was a bug that sample identifier didn't match actual identifiers. Hypothetically, it would have been
less backwards-incompat to make sample identifier use slash. But let's use 1.0 to make this conform to oai-pmh guidance.

Especially becuase the majority of current users of this gem are probably using it via blacklight_oai_provider, and it was already providing runtime patches to the oai gem to get colon behavior, and had a commit message suggesting it believed it was patching an "identifier bug in oai gem." So let's fix the thing our main user thinks was a bug!
https://github.com/projectblacklight/blacklight_oai_provider/blob/f4b25ee43846cb9efc9d5124a8639a496153aa18/config/initializers/oai_patches.rb

Closes #38
Ref #61