Support testing on Rails 6.1 and 7.0
We don't have CI working on this repo (See #87), but I manually ran tests in Rails 5.2, 6.0. 6.1, and 7.0 -- it passed on all of them with no changes

BUT there was a deprecation notice in 7. I went ahead and fixed the deprecation notice (also confirming by manually running tests that still passed on all those versions).

This will at least allow manual testing on more recent Rails versions, as well as set the path for testing on recent Rails when we restore CI.