If AR model provides no "sets", then default should be [] instead of nil to avoid uncaught exception
Previously, if the model defined no sets, and a client asked for them, an uncaught exception would be raised, `NoMethodError: undefined method `detect' for nil`

By making the default return `[]` instead of `nil`, it changes to an appropriate `OAI::NoMatchException` which is rescued properly by controller etc.

I do not use sets, and am not totally sure functionality is working if you do, haven't tracked down a list. But pretty sure this wont' break anything.